// moonfowcl/src/app/model/common/embedded-options.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

import {EmbeddedOptionsIF} from './embedded-options-if';

export class EmbeddedOptions {

  urlEsciModulo: string;
  urlTornaIstanze: string;
  urlErrore: string;
  tabIstanze: string[];

//   constructor(embeddeOptionsIF: EmbeddedOptionsIF) {
//     this.urlEsciModulo = embeddeOptionsIF.urlEsciModulo;
//     this.urlTornaIstanze = embeddeOptionsIF.urlTornaIstanze;
//     this.urlErrore = embeddeOptionsIF.urlErrore;
//     this.tabIstanze = embeddeOptionsIF.tabIstanze;
//  }

}